Modify: 상품 테이블 컴포넌트화
